Add DataUnit interface
Let this be an abstraction that represents one dimension of data (i.e., a
feature or sample) and have DataFeature and DataSample extend it. This is
needed to facilitate creating Task plans, since the algorithm needs to evaluate
which tasks have common work, apathetic to whether it's common features or
samples.