Optimize task plan algorithm
Finds tasks by enumerating through the set of all work present in the job group
and creating tasks for all jobs that contain that work. This makes finding
common tasks O(mn) rather than O(2^n), which is much much better.

Also changes the search algorithm to a sort of best-first search with a
priority queue that uses the distance away from a terminating task as a
heuristic (since we were already keeping track of it). This makes it much more
scalable for more work and jobs as well.

Note that a job group should ideally be very correlated beforehand, otherwise
the algorithm can still take a long time when given a lot of jobs that only
have a few tasks in common with one another.