Make task plan depth-first and find subgroups
Before, the task plan algorithm was working breadth-first, which was causing it
to run out of memory and otherwise take forever for large sets of jobs that had
little work in common.

Secondly, the task plan finds subgroups before running the algorithm, making
sure each sub-plan starts with a task that has work shared by *all* jobs in the
tree. This is an underlying assumption in the algorithm, and before we were
artificially bloating the distance heuristic by saying that traversing tasks
associated with work not in the sub-plan still costed a step.